Remove interaction v1 views and related code
